requirements.txt
asgiref==3.8.1
Django==5.1.3
django-allauth==65.2.0
django-environ==0.11.2
djangorestframework==3.15.2
numpy==2.1.3
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
